fix(tui): avoid duplicate fallback notice
