+++ packages/coding-agent/src/modes/interactive/components/assistant-message.ts
@@ -1,9 +1,8 @@
 import type { AssistantMessage } from "@earendil-works/pi-ai";
 import { type Component, Container, Markdown, type MarkdownTheme, Spacer, Text } from "@earendil-works/pi-tui";
 import type { MarkdownTransformer } from "../../../core/extensions/types.ts";
-import { formatDuration } from "../../../utils/duration.ts";
-import { formatProviderNativeBody, formatProviderNativeSummary } from "../../provider-native-rendering.ts";
 import { getMarkdownTheme, theme } from "../theme/theme.ts";
+import { type AssistantRenderDescriptor, createAssistantRenderDescriptors } from "./assistant-render-descriptors.ts";
 import { createMarkdownTransform } from "./markdown-transform.ts";
 import { createBoundedRenderSignature } from "./render-signature.ts";
 
@@ -11,30 +10,8 @@ const OSC133_ZONE_START = "\x1b]133;A\x07";
 const OSC133_ZONE_END = "\x1b]133;B\x07";
 const OSC133_ZONE_FINAL = "\x1b]133;C\x07";
 
-type MarkdownDescriptorKind = "text-md" | "thinking-md";
-type TextDescriptorKind = "thinking-label" | "provider-native-summary" | "provider-native-body" | "error-text";
-type RenderDescriptorKind = "spacer" | MarkdownDescriptorKind | TextDescriptorKind;
-type RenderDescriptor = { readonly kind: RenderDescriptorKind; readonly text: string };
-
-const SPACER_DESCRIPTOR = { kind: "spacer", text: "" } as const satisfies RenderDescriptor;
-
 function assertNever(value: never): never {
 	throw new TypeError(`Unexpected assistant render variant: ${String(value)}`);
-}
-
-function isVisibleContent(content: AssistantMessage["content"][number], providerNativeVisible: boolean): boolean {
-	switch (content.type) {
-		case "text":
-			return Boolean(content.text.trim());
-		case "thinking":
-			return Boolean(content.thinking.trim());
-		case "providerNative":
-			return providerNativeVisible;
-		case "toolCall":
-			return false;
-		default:
-			return assertNever(content);
-	}
 }
 
 export class AssistantMessageComponent extends Container {
@@ -47,7 +24,7 @@ export class AssistantMessageComponent extends Container {
 	private markdownTransformers: readonly MarkdownTransformer[];
 	private lastMessage?: AssistantMessage;
 	private lastMessageSignature?: string;
-	private renderDescriptors: readonly RenderDescriptor[] = [];
+	private renderDescriptors: readonly AssistantRenderDescriptor[] = [];
 	private hasToolCalls = false;
 	private expanded = false;
 	private isStreaming = false;
@@ -137,122 +114,16 @@ export class AssistantMessageComponent extends Container {
 		this.renderCache = undefined;
 		if (streamingChanged) this.renderDescriptors = [];
 		this.hasToolCalls = message.content.some((content) => content.type === "toolCall");
-		const descriptors = this.createRenderDescriptors(message);
+		const descriptors = createAssistantRenderDescriptors(message, {
+			expanded: this.expanded,
+			hiddenThinkingLabel: this.hiddenThinkingLabel,
+			hideThinkingBlock: this.hideThinkingBlock,
+			hasToolCalls: this.hasToolCalls,
+		});
 		this.reconcileRenderDescriptors(descriptors);
 	}
 
-	private createRenderDescriptors(message: AssistantMessage): readonly RenderDescriptor[] {
-		const descriptors: RenderDescriptor[] = [];
-		if (message.content.some((content) => isVisibleContent(content, true))) descriptors.push(SPACER_DESCRIPTOR);
-		for (let i = 0; i < message.content.length; i++) {
-			const content = message.content[i];
-			switch (content.type) {
-				case "text": {
-					const text = content.text.trim();
-					if (text) descriptors.push({ kind: "text-md", text });
-					break;
-				}
-				case "thinking": {
-					const thinkingBlocks: string[] = [];
-					let hasTiming = false;
-					let isDone = true;
-					let minStart = Number.POSITIVE_INFINITY;
-					let maxEnd = Number.NEGATIVE_INFINITY;
-					for (; i < message.content.length; i++) {
-						const thinkingContent = message.content[i];
-						if (thinkingContent.type !== "thinking") break;
-						const startedAt = thinkingContent.startedAt;
-						if (startedAt !== undefined) {
-							hasTiming = true;
-							minStart = Math.min(minStart, startedAt);
-							const endedAt = thinkingContent.endedAt;
-							if (endedAt === undefined) {
-								isDone = false;
-							} else {
-								maxEnd = Math.max(maxEnd, endedAt);
-							}
-						}
-						const thinking = thinkingContent.thinking.trim();
-						if (thinking) thinkingBlocks.push(thinking);
-					}
-					i--;
-					if (thinkingBlocks.length === 0) break;
-					if (!hasTiming) {
-						const text = this.hideThinkingBlock
-							? theme.italic(theme.fg("thinkingText", this.hiddenThinkingLabel))
-							: thinkingBlocks.join("\n\n");
-						descriptors.push({ kind: this.hideThinkingBlock ? "thinking-label" : "thinking-md", text });
-					} else {
-						const label = isDone
-							? theme.italic(
-									theme.fg("thinkingText", `Thought: ${formatDuration(Math.max(0, maxEnd - minStart))}`),
-								)
-							: theme.italic(theme.fg("thinkingText", this.hiddenThinkingLabel));
-						if (this.hideThinkingBlock) {
-							descriptors.push({ kind: "thinking-label", text: label });
-						} else {
-							descriptors.push(
-								{ kind: "thinking-label", text: label },
-								{ kind: "thinking-md", text: thinkingBlocks.join("\n\n") },
-							);
-						}
-					}
-					if (message.content.slice(i + 1).some((following) => isVisibleContent(following, false)))
-						descriptors.push(SPACER_DESCRIPTOR);
-					break;
-				}
-				case "providerNative":
-					descriptors.push(
-						{
-							kind: "provider-native-summary",
-							text: theme.fg("muted", formatProviderNativeSummary(message, content, this.expanded)),
-						},
-						{
-							kind: "provider-native-body",
-							text: theme.fg("dim", formatProviderNativeBody(content, this.expanded)),
-						},
-					);
-					if (message.content.slice(i + 1).some((following) => isVisibleContent(following, true)))
-						descriptors.push(SPACER_DESCRIPTOR);
-					break;
-				case "toolCall":
-					break;
-				default:
-					assertNever(content);
-			}
-		}
-		const addError = (text: string): void => {
-			descriptors.push(SPACER_DESCRIPTOR, { kind: "error-text", text: theme.fg("error", text) });
-		};
-		switch (message.stopReason) {
-			case "length":
-				addError(
-					"Error: Model stopped because it reached the maximum output token limit. The response may be incomplete.",
-				);
-				break;
-			case "aborted": {
-				if (this.hasToolCalls) break;
-				const abortMessage =
-					message.errorMessage && message.errorMessage !== "Request was aborted"
-						? message.errorMessage
-						: "Operation aborted";
-				addError(abortMessage);
-				break;
-			}
-			case "error":
-				if (!this.hasToolCalls) addError(`Error: ${message.errorMessage || "Unknown error"}`);
-				break;
-			case "pending":
-			case "stop":
-			case "toolUse":
-				break;
-			default:
-				assertNever(message.stopReason);
-		}
-		return descriptors;
-	}
-
-	private reconcileRenderDescriptors(descriptors: readonly RenderDescriptor[]): void {
+	private reconcileRenderDescriptors(descriptors: readonly AssistantRenderDescriptor[]): void {
 		let divergentIndex = 0;
 		const sharedLength = Math.min(this.renderDescriptors.length, descriptors.length);
 		while (divergentIndex < sharedLength) {
@@ -273,7 +144,7 @@ export class AssistantMessageComponent extends Container {
 		this.renderDescriptors = descriptors;
 	}
 
-	private createRenderChild(descriptor: RenderDescriptor): Component {
+	private createRenderChild(descriptor: AssistantRenderDescriptor): Component {
 		switch (descriptor.kind) {
 			case "spacer":
 				return new Spacer(1);
@@ -312,7 +183,7 @@ export class AssistantMessageComponent extends Container {
 			content: message.content,
 			hiddenThinkingLabel: this.hiddenThinkingLabel,
 			hideThinkingBlock: this.hideThinkingBlock,
-			errorMessage: message.errorMessage,
+			errorState: [message.diagnostics, message.errorMessage],
 			stopReason: message.stopReason,
 		});
 	}
